Clean up Grunt configuration.
