+++ lib/themes/dosomething/paraneue_dosomething/Gruntfile.js
@@ -1,43 +1,20 @@
 /* jshint node:true */
 "use strict";
 
-module.exports = function(grunt) {
-  // @TODO: Remove this in a bit.
-  var sass = require('node-sass');
-  var fs = require('fs');
+var path = require('path');
 
-  // measures the time each task takes
+module.exports = function(grunt) {
+  // Measure the time each task takes
   require('time-grunt')(grunt);
 
-  // Load tasks from `/tasks`
+  // Load tasks from `tasks/`
   grunt.loadTasks('tasks');
 
-  function writeFile(path, contents) {
-    if(grunt.file.exists(path)) {
-      grunt.file.delete(path);
-    }
-
-    grunt.file.write(path, contents);
-  }
-
-  grunt.registerTask('nodesass', function() {
-    console.log(sass.info());
-
-    var result = sass.renderSync({
-      file: __dirname + '/scss/app.scss',
-      outFile: __dirname + '/dist/app.min.css',
-      outputStyle: 'compressed',
-      sourceMap: false
-    });
-
-    writeFile(__dirname + '/dist/app.min.css', result.css);
-  });
-
-  // Load plugin configuration from `tasks/options`.
+  // Load plugin configuration from `tasks/`.
   require('load-grunt-config')(grunt, {
-    configPath: require('path').join(process.cwd(), 'tasks/options'),
+    configPath: path.join(process.cwd(), 'tasks'),
     init: true,
-    config: {
+    data: {
       pkg: grunt.file.readJSON("package.json")
     }
   });
