+++ lib/themes/dosomething/paraneue_dosomething/tasks/aliases.json
@@ -1,0 +1,5 @@
+{
+  "default": ["lint", "build", "watch"],
+  "build": ["clean:dist", "copy:assets", "sass", "postcss", "requirejs", "modernizr"],
+  "lint": ["jshint", "scsslint"]
+}
