+++ lib/themes/dosomething/paraneue_dosomething/tasks/clean.js
@@ -1,3 +1,3 @@
 module.exports = {
   dist: ["dist/"]
-}
+};
