+++ lib/themes/dosomething/paraneue_dosomething/tasks/copy.js
@@ -4,4 +4,4 @@ module.exports = {
       {expand: true, src: ["images/**"], dest: "dist/"}
     ]
   }
-}
+};
