--- lib/themes/dosomething/paraneue_dosomething/tasks/default.js
@@ -1,3 +1,0 @@
-module.exports = function(grunt) {
-  grunt.registerTask("default", ["lint", "build", "watch"]);
-};
