--- lib/themes/dosomething/paraneue_dosomething/tasks/lint.js
@@ -1,6 +1,0 @@
-module.exports = function(grunt) {
-  grunt.registerTask("lint", ["lint:js", "lint:css"]);
-
-  grunt.registerTask("lint:js", ["jshint:all"]);
-  grunt.registerTask("lint:css", ["scsslint:scss"]);
-}
