+++ lib/themes/dosomething/paraneue_dosomething/tasks/modernizr.js
@@ -1,5 +1,5 @@
 module.exports = {
-  dist: {
+  all: {
     "devFile": "remote",
     "outputFile": "dist/modernizr.js",
     "files" : {
