+++ lib/themes/dosomething/paraneue_dosomething/tasks/postcss.js
@@ -1,5 +1,5 @@
 module.exports = {
-  process: {
+  all: {
     src: "dist/app.min.css",
     options: {
       map: false,
