+++ lib/themes/dosomething/paraneue_dosomething/tasks/scsslint.js
@@ -1,5 +1,5 @@
 module.exports = {
-  scss: [
+  all: [
     'scss/**/*.scss'
   ],
