+++ lib/themes/dosomething/paraneue_dosomething/tasks/watch.js
@@ -1,14 +1,14 @@
 module.exports = {
   sass: {
     files: ["scss/**/*.{scss,sass}"],
-    tasks: ["lint:css", "nodesass",  "postcss:process"]
+    tasks: ["scsslint", "sass",  "postcss"]
   },
   js: {
     files: ["js/**/*.js", "tests/**/*.js"],
-    tasks: ["lint:js", "requirejs:dev"]
+    tasks: ["jshint", "requirejs"]
   },
   assets: {
     files: ["assets/**/*", "bower_components/**/*"],
     tasks: ["copy:assets"]
   }
-}
+};
